// src/Core/interfaces/col-meter-table.interface.ts
import { NzTableFilterFn, NzTableFilterList, NzTableSortFn, NzTableSortOrder } from 'ng-zorro-antd/table';

export interface ColumnItem {
    name: string;
    sortOrder: NzTableSortOrder | null;
    sortFn: NzTableSortFn<any> | null;
    listOfFilter: NzTableFilterList;
    filterFn: NzTableFilterFn<any> | null;
    filterMultiple: boolean;
    sortDirections: NzTableSortOrder[];
  }